// src/modules/safety/enums/safety-incident-type.enum.ts
export enum SafetyIncidentType {
  MEDICAL_EMERGENCY = 'MEDICAL_EMERGENCY',
  ACCIDENT = 'ACCIDENT',
  THREAT = 'THREAT',
  HARASSMENT = 'HARASSMENT',
  ROBBERY = 'ROBBERY',
  UNSAFE_DRIVING = 'UNSAFE_DRIVING',
  VEHICLE_FAILURE = 'VEHICLE_FAILURE',
  LOST_CONTACT = 'LOST_CONTACT',
  OTHER = 'OTHER',
}